Some cleaning in Aggregate class
