+++ quantarhei/builders/aggregate_spectroscopy.py
@@ -1,0 +1,7 @@
+# -*- coding: utf-8 -*-
+
+from .aggregate_base import AggregateBase
+
+
+class AggregateSpectroscopy(AggregateBase):
+    pass
